Crate pallet_bridge_beefy

BEEFY bridge pallet.

This pallet is an on-chain BEEFY light client for Substrate-based chains that are using the following pallets bundle: pallet-mmr, pallet-beefy and pallet-beefy-mmr.

The pallet is able to verify MMR leaf proofs and BEEFY commitments, so it has access to the following data of the bridged chain:

header hashes

changes of BEEFY authorities

extra data of MMR leafs

Given the header hash, other pallets are able to verify header-based proofs (e.g. storage proofs, transaction inclusion proofs, etc.).

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

ImportedCommitmentsInfoData

Some high level info about the imported commitments.

Constants

LOG_TARGET

The target that will be used when publishing logs related to this pallet.

Type Aliases

BridgedBeefyAuthorityId

BEEFY validator id, used by configured bridged chain.

BridgedBeefyAuthoritySet

BEEFY validator set, used by configured bridged chain.

BridgedBeefyAuthoritySetInfo

BEEFY authority set, used by configured bridged chain.

BridgedBeefyCommitmentHasher

BEEFY commitment hasher, used by configured bridged chain.

BridgedBeefyMmrLeaf

MMR leaf type, used by configured bridged chain.

BridgedBeefyMmrLeafExtra

The type of the MMR leaf extra data used by the configured bridged chain.

BridgedBeefySignedCommitment

BEEFY signed commitment, used by configured bridged chain.

BridgedBlockHash

Block hash, used by configured bridged chain.

BridgedBlockNumber

Block number, used by configured bridged chain.

BridgedChain

Configured bridged chain.

BridgedMmrHash

MMR hashing output type of BridgedMmrHashing<T, I>.

BridgedMmrHashing

MMR hashing algorithm, used by configured bridged chain.

BridgedMmrProof

BEEFY MMR proof type used by the pallet

ImportedCommitment

Imported commitment data, stored by the pallet.

InitializationDataOf

Pallet initialization data.