Trait pallet_bridge_beefy::pallet::Config

pub trait Config<I: 'static = ()>: Config {
    type MaxRequests: Get<u32>;
    type CommitmentsToKeep: Get<u32>;
    type BridgedChain: ChainWithBeefy;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type MaxRequests: Get<u32>

The upper bound on the number of requests allowed by the pallet.

A request refers to an action which writes a header to storage.

Once this bound is reached the pallet will reject all commitments until the request count has decreased.

type CommitmentsToKeep: Get<u32>

Maximal number of imported commitments to keep in the storage.

The setting is there to prevent growing the on-chain state indefinitely. Note the setting does not relate to block numbers - we will simply keep as much items in the storage, so it doesn’t guarantee any fixed timeframe for imported commitments.

type BridgedChain: ChainWithBeefy

The chain we are bridging to here.

Object Safety

This trait is not object safe.