pallet_bridge_beefy

Type Alias BridgedBeefyMmrLeaf

pub type BridgedBeefyMmrLeaf<T, I> = BeefyMmrLeafOf<BridgedChain<T, I>>;

MMR leaf type, used by configured bridged chain.

Aliased Type

struct BridgedBeefyMmrLeaf<T, I> {
    pub version: MmrLeafVersion,
    pub parent_number_and_hash: (<<T as Config<I>>::BridgedChain as Chain>::BlockNumber, <<T as Config<I>>::BridgedChain as Chain>::Hash),
    pub beefy_next_authority_set: BeefyAuthoritySet<<<T as Config<I>>::BridgedChain as ChainWithBeefy>::MmrHash>,
    pub leaf_extra: <<T as Config<I>>::BridgedChain as ChainWithBeefy>::BeefyMmrLeafExtra,
}

Fields
version: MmrLeafVersion

Version of the leaf format.

Can be used to enable future format migrations and compatibility. See [MmrLeafVersion] documentation for details.

parent_number_and_hash: (<<T as Config<I>>::BridgedChain as Chain>::BlockNumber, <<T as Config<I>>::BridgedChain as Chain>::Hash)

Current block parent number and hash.

beefy_next_authority_set: BeefyAuthoritySet<<<T as Config<I>>::BridgedChain as ChainWithBeefy>::MmrHash>

A merkle root of the next BEEFY authority set.

leaf_extra: <<T as Config<I>>::BridgedChain as ChainWithBeefy>::BeefyMmrLeafExtra

Arbitrary extra leaf data to be used by downstream pallets to include custom data in the [MmrLeaf]